.travis.yml: fix coverage generation line.
Using the `--enable-coverage` switch in mamba seems to exercise a lot of files
besides the pysellus package, so its measurement is wrong.

We can just use the `coverage` package instead, which is a dependency of mamba,
without setting that switch. The `script` line in .travis.yml now runs the test
and fails if they fail. After success, the tests are re-run and coverage is
assessed, then reported to coveralls.io.

Reporting to coveralls should work fine after this.