Change Threader to work with Subjects
Threader used to create as many threads as the sum of all the function
lists in the given observable map. This worked ok for a quick prototype
but could get ugly when a single stream has a lot of functions applied
to it.

Adding subjects, we cut the number of threads to a minimum, that is, one
thread per stream in the map.

Instead of creating a single thread per function, we subscribe each
function to a subject (proxy), and then create a thread per subject.

I modified the tests too to account for the change in the number of
threads too.

Potential problems:

	- Threader now knows about rx.subjects. This could be changed to
	  make threader same as before and make registrar work with them.
	  But this just shifts the problem over to registrar, and makes
	  the structure of `stream_to_observers` more complex.