Add option to launch pysellus from a single file
New options allows to test the program with a single file without having
to rename all files in the parent directory to __file.

Loader now tests if the given path is a python file (ends with .py).
If it does, import it and get all pscheck functions from it in the same
way we did with directories.

To avoid duplication, the process of loading the check functions from a
module has been moved to its own function: `_get_checks_from_module`

Tests have been modified to add a new test case for the new
functionality, and changed old descriptions.