Add on_failure decorator
In our original spec, the failure message provided to the integration
had to tell the test it failed on. This proved very difficult to
implement with the model we chose on the first place: that is, attach
@on_failure to the tester functions.

With this change, we approach integrations differently. Intead of
decorating the tester functions, we decorate the `pscheck_` function
that we transpile from the `@check` DSL functions. This allows us to get
the function name, and allows the user to choose where to notify him/her
depending on the check, not the test.

For this to work, we had to modify the original implementation of
expect. Before, it only mapped the given stream to a list of
functions.

@on_failure('terminal')
def pscheck_function():
    ...
    expect(foo)(bar)

First, @on_failure maps the pscheck function to a list of integrations
(implemented internally as Rx Subjects):

Then, expect takes a stream (Rx Observable), and returns a registrar
function, that takes a variable number of tester functions. For each of
those, defines a wrapper function and registers that, mapping pscheck
functions with a list of tests used inside its body.

The wrapper function takes an element, and passes it to the original
tester. When the tester fails, it will notify all integrations passed to
the on_failure decorator, passing the element where it failed, the
`pscheck` function where it happened, and the tester function name. It
will also pass a brief description of the problem.

In addition, if, for whatever reason, an exception is raised inside the
tester, we will catch it and notify the integrations of it, with a
different message that will include the exception message.