Add integration loader
Adds `integration_config` module, that, on start-up, goes and looks for
a configuration file in the supplied directory (for now, the config file
name is `.ps_integrations.yml`.

That configuration file will define user aliases to stock or custom
integrations, and supply some constructor arguments for them (like API
keys, or something else). `integration_config` will go ahead and
instance all classes with those parameters and store it in a dictionary
mapping user aliases to instances.

Then, `integrations#_create` will look in that dictionary, and get the
subject for those classes.

If any error occurs, print an error and exit the program