Notify integrations of function docstring
Previously, we were sending the function name to the integration, but
this didn't make much sense as the user never sees this name when
writing tests.

With past changes, we removed the condition of prepending `pscheck_` to
the tester functions, and instead add an attribute to it. Now, we remove
the limitation of having to write long, descriptive function names, and
instead focus on an user-written docstring in the test function.

This paves the way for an easier DSL processing when parsing, we add the
user test description as a docstring, and then send that to the
integration, so the user never knows the intermediate expanded function
name.