Remove super().__init__() requirement from Integrations
Previously, AbstractIntegration would create a subject on __init__, that
would get returned when calling get_subject. This made all built-in and
user-defined integrations call the super constructor if they had to
define a custom constructor.

Now, the subject is defined, instantiaded and subscribed on the same
get_subject super method. In theory, this method only gets called once,
so this shouldn't cause any problems, but we'll see.