Support custom integrations
The syntax to add a new custom integration is as follows:

```
custom_integrations:
    your_integration_name:
	path: "/absolute/path/to/integration/file"
	name: "NameOfIntegrationClass"
```

For each custom integration, the user should supply the path of the
integration file, and the name of the class containing the integration
inside the file (that would be the class inheriting from
AbstractIntegration).

The program will abort if no matching class is found inside the given
file.