integration_config_spec: amend the purpose of some integration tests
The two tests verifying the behaviour when the config file
doesn't exist were passing because the bogus paths obviously
made the program raise FileNotFoundError. But we wanted to check
that, directories that actually exist, but do not contain a config
file, that exception is raised. This patch changes those tests
to do just that.

This patch also refactors integration_config, extracting a method
from `_load_config_file`.